requirements.txt
# Core application dependencies
asyncio-mqtt==0.16.1
beautifulsoup4==4.12.3
fastapi==0.109.1
fastapi-websocket==0.1.7
jinja2==3.1.3
openai==1.54.0
pydantic==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
uvicorn==0.27.0
websockets==12.0

# Audio processing
sounddevice>=0.4.6
scipy>=1.11.0
librosa>=0.10.0
numpy>=1.24.0

# TUI and visualization
textual>=0.41.0
rich>=13.0.0

# Optional audio backends
pyaudio==0.2.13

# Development and testing
pytest==8.0.0
pytest-asyncio==0.23.5

# Additional utilities for audio format support
soundfile>=0.12.1  # For advanced audio file format support

# Function calling and AI integration
anthropic>=0.35.0  # For Claude integration if needed 